requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
moviepy>=1.0.3
python-dateutil>=2.8.2
